+++ project-euler-rust/src/registry.rs
@@ -1,6 +1,6 @@
 pub struct SolutionEntry {
-    pub id: u32,
-    pub solve: fn() -> i32,
+    pub id: i32,
+    pub solve: fn() -> i64,
 }
 
 inventory::collect!(SolutionEntry);
